chore:move message modal component to modal folder
