+++ src/components/modal/MessageModal.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Modal from '../../components/modal/Modal';
+import Modal from './Modal';
 
 export default function MessageModal(props) {
   return (
